Added playing an audio on interacting with an interactable.
The demo is set up for "Activation" which is pressing the trigger after an object is held, like firing a gun. In many cases, this will want to be changed to the "selection Enter" interaction event, where its just when you pick up to investigate an object, where something happens.