Set up example to have room switched via timeline
Lightgroup members can be set up in scene with default intensity, and cut to 0 with an option.
signal receivers and emitters are all that's needed to connect to various objects in the scene, with unity's timeline.
timelines could be set up as short, just for transitions, and triggered elsewhere in code/interaction.

bug fixes in GameObjectScene, plus a jump to index function.